+++ test/bootstrap.php
@@ -1,4 +1,0 @@
-<?php
-
-// Archer uses Asplode for error handling if it's available, but we don't want that
-restore_error_handler();