+++ test/suite/Eloquent/Asplode/FunctionalTest.php
@@ -11,10 +11,46 @@
 
 class FunctionalTest extends PHPUnit_Framework_TestCase
 {
+    protected function setUp()
+    {
+        parent::setUp();
+
+        $this->handlerStack = \Eloquent\Asplode\Asplode::removeErrorHandlers();
+    }
+
+    protected function tearDown()
+    {
+        parent::tearDown();
+
+        \Eloquent\Asplode\Asplode::removeErrorHandlers();
+        \Eloquent\Asplode\Asplode::restoreErrorHandlers($this->handlerStack);
+    }
+
     /**
      * Test one line installation.
      */
     public function testOneLineInstallation()
+    {
+        if (!version_compare(PHP_VERSION, '5.4')) {
+            $this->markTestSkipped('Requires PHP >= 5.4');
+        }
+
+        (new \Eloquent\Asplode\Asplode)->install();
+        $actual = set_error_handler(function() {
+        });
+        restore_error_handler();
+        restore_error_handler();
+
+        $this->assertInstanceOf(
+            '\Eloquent\Asplode\Asplode',
+            $actual
+        );
+    }
+
+    /**
+     * Test one line installation for PHP 5.3.
+     */
+    public function testOneLineInstallationPhp53()
     {
         \Eloquent\Asplode\Asplode::instance()->install();
         $actual = set_error_handler(function() {
@@ -22,20 +58,10 @@ class FunctionalTest extends PHPUnit_Framework_TestCase
         restore_error_handler();
         restore_error_handler();
 
-        $this->assertTrue(is_array($actual));
-        $this->assertSame(array(0, 1), array_keys($actual));
-        $this->assertInstanceOf('Eloquent\Asplode\Asplode', $actual[0]);
-        $this->assertSame('handleError', $actual[1]);
-        $this->assertTrue(is_callable($actual));
-    }
-
-    /**
-     * Test legacy PHP error example.
-     */
-    public function testLegacyPhpError()
-    {
-        $this->setExpectedException('PHPUnit_Framework_Error_Warning');
-        $fp = fopen(uniqid(), 'r');
+        $this->assertInstanceOf(
+            '\Eloquent\Asplode\Asplode',
+            $actual
+        );
     }
 
     /**